usertests bug. do not know why fuck

+++ kernel/buf.h
@@ -7,7 +7,7 @@ struct buf {
   uint refcnt;
   // struct buf *prev; // LRU cache list
   struct buf *next;
-  uint ticks;
+  uint64 ticks;
   uchar data[BSIZE];
 };
